squashed and cherry-picked from #1104
[wip] use transaction API

rename

add transaction bindings
