+++ python/aiertx.py
@@ -1,0 +1,5 @@
+# Copyright (C) 2022, Advanced Micro Devices, Inc.
+# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+
+# noinspection PyUnresolvedReferences
+from ._mlir_libs._aiertx import *
